Restructure coral-embed-stream
