+++ client/coral-embed-stream/src/tabs/stream/components/EditableCommentContent.js
