+++ client/coral-embed-stream/src/tabs/stream/components/Stream.css
