+++ client/coral-embed-stream/src/tabs/stream/components/StreamError.css
