+++ client/coral-embed-stream/src/tabs/stream/components/SuspendedAccount.css
