+++ client/coral-embed-stream/src/tabs/stream/containers/AutomaticAssetClosure.js
